test: improve benchmark + adapt tests for huge images (16,5GB)
The test_cuda framework wasn't prepared to deal with huge images. There are still issues with the read_input() function to be solved, they are causing memory failiure.

+++ src/chunkedExecutor.cu
@@ -331,7 +331,7 @@ void chunkedExecutorFillHoles(Func func, int ncopies, const float safetyMargin, 
                               Args... args) {
 
   // set output initial data
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(dtype);
 
   dtype* margins = (dtype*)malloc(nBytes);
