+++ src/morphology/binary/smooth_binary.cu
@@ -129,7 +129,7 @@ void smooth_binary_on_host(dtype* hostImage, dtype* hostOutput, const int xsize,
                            int kernel_zsize) {
 
   // set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(dtype);
 
   // allocate temporary memory
