+++ src/morphology/grayscale/morph_grayscale.cu
@@ -167,7 +167,7 @@ void morph_grayscale_on_device(dtype* hostImage, dtype* hostOutput, const int xs
                                int kernel_xsize, int kernel_ysize, int kernel_zsize,
                                MorphOp operation) {
   // set input dimension
-  size_t size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(dtype);
   size_t nBytes_padding = xsize * ysize * (padding_bottom + padding_top) * sizeof(dtype);
   size_t nBytes_input = nBytes + nBytes_padding;
