+++ tests_cuda/morphology/binary/test_geodesic_morph_binary.cpp
@@ -35,11 +35,11 @@ void test_geodesic_morph_binary_on_device(const std::string& filename, const int
   printf("\nTest binary geodesic %s on device\n", (operation ? "dilation" : "erosion"));
 
   // Set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(int);
 
   if (flag_verbose)
-    printf("Matrix size: %d (%d.%d.%d)\n", size, xsize, ysize, zsize);
+    printf("Matrix size: %zu (%d.%d.%d)\n", size, xsize, ysize, zsize);
 
   int *hostMarker, *device_ref, *hostMask;  // Pointers for host memory
   hostMarker = (int*)malloc(nBytes);
@@ -114,11 +114,11 @@ void test_geodesic_morph_binary_on_host(const std::string& filename, const int x
   printf("\nTest binary geodesic %s on host\n", (operation ? "dilation" : "erosion"));
 
   // Set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(int);
 
   if (flag_verbose)
-    printf("Matrix size: %d (%d.%d.%d)\n", size, xsize, ysize, zsize);
+    printf("Matrix size: %zu (%d.%d.%d)\n", size, xsize, ysize, zsize);
 
   int *hostMarker, *host_ref, *hostMask;  // Pointers for host memory
   hostMarker = (int*)malloc(nBytes);
