+++ tests_cuda/morphology/binary/test_morph_binary.cpp
@@ -34,11 +34,11 @@ void test_morph_binary_on_device(const std::string& filename, const int xsize, c
   printf("\nTest binary %s on device\n", (operation ? "dilation" : "erosion"));
 
   // Set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(int);
 
   if (flag_verbose)
-    printf("Matrix size: %d (%d.%d.%d)\n", size, xsize, ysize, zsize);
+    printf("Matrix size: %zu (%d.%d.%d)\n", size, xsize, ysize, zsize);
 
   int *host_A, *device_ref;  // Pointers for host memory
   host_A = (int*)malloc(nBytes);
@@ -95,11 +95,11 @@ void test_morph_binary_on_host(const std::string& filename, const int xsize, con
   printf("\nTest binary %s on host\n", (operation ? "dilation" : "erosion"));
 
   // Set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
   size_t nBytes = size * sizeof(int);
 
   if (flag_verbose)
-    printf("Matrix size: %d (%d.%d.%d)\n", size, xsize, ysize, zsize);
+    printf("Matrix size: %zu (%d.%d.%d)\n", size, xsize, ysize, zsize);
 
   int *host_A, *host_ref;  // Pointers for host memory
   host_A = (int*)malloc(nBytes);
