+++ tests_cuda/morphology/grayscale/test_bottom_hat.cpp
@@ -17,12 +17,12 @@ void test_bottom_hat_on_device(const std::string& filename, const int xsize, con
   printf("\nTest bottom hat on device\n");
 
   // set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
 
   size_t nBytes = size * sizeof(float);
 
   if (flag_verbose)
-    printf("Matrix size:   %d (%d.%d.%d)\n", size, xsize, ysize, zsize);
+    printf("Matrix size:   %zu (%d.%d.%d)\n", size, xsize, ysize, zsize);
 
   float *host_A, *device_ref;  //pointers for host memory
   host_A = (float*)malloc(nBytes);
@@ -66,11 +66,11 @@ void test_bottom_hat_on_host(const std::string& filename, const int xsize, const
   printf("\nTest bottom hat on host\n");
 
   // set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = static_cast<size_t>(xsize) * static_cast<size_t>(ysize) * static_cast<size_t>(zsize);
 
   size_t nBytes = size * sizeof(float);
   if (flag_verbose)
-    printf("Matrix size:   %d (%d.%d.%d)\n", size, xsize, ysize, zsize);
+    printf("Matrix size:   %zu (%d.%d.%d)\n", size, xsize, ysize, zsize);
 
   float *host_A, *host_ref;  //pointers for host memory
   host_A = (float*)malloc(nBytes);
